Display loaded files for ctu analysis (#199)
During CTU, if a function definition is needed from another TU and the
pch dump is present we display the name of the source file from which
the dump is generated (.ast extension removed).
This is needed to be able to collect those TUs which are used during the
analysis; in case of a crash we can package these files into the failure
zip.