Fix redecl chain of classes and class templates
The crux of the issue was that lookup could not find previous decls of
a friend class. The solution involves making the friend declarations
visible in their decl context (i.e. adding them to the lookup table).
This fix involves two other repairs:
(1) We could not handle the addition of injected class types properly
when a redecl chain was involved, now this is fixed.
(2) DeclContext::removeDecl failed if the lookup table in Vector form
did not contain the to be removed element. This caused troubles in
ASTImporter::ImportDeclContext. This is also fixed.