Add CTU On-Demand analysis support
Add an option to enable on-demand parsing of needed ASTs during CTU
analysis, and another option to specify the compilation database used.
The option CTUOnDemandParsing is a boolean flag, which enables the new
AST-loading mode. Option CTUOnDemandParsingDatabase is a string, which
should be the path of the compilation database used for on-demand parsing.
The compilation database is needed for on-demand mode, because it has all the
necessary information to generate the ASTs.

Please enter the commit message for your changes. Lines starting